restructered folder structure, moved email and auth into utils folder
