--- backend/util/errorUtil.js
@@ -1,7 +1,0 @@
-const createError = (statusCode, errorMessage) => {
-    const error = new Error(errorMessage);
-    error.statusCode = statusCode;
-    return error;
-}
-
-export default {createError}
